Move API\Photo testcases
